Move color composites controller to module file
